Move spark-cassandra to backend module
