+++ it/src/test/scala/quasar/TestConfig.scala
@@ -53,8 +53,7 @@ object TestConfig {
   val SPARK_HDFS      = ExternalBackendRef(BackendRef(BackendName("spark_hdfs")       , BackendCapability.All), sparkcore.fs.hdfs.SparkHdfsBackendModule.Type)
   val SPARK_LOCAL     = ExternalBackendRef(BackendRef(BackendName("spark_local")      , BackendCapability.All), sparkcore.fs.local.SparkLocalBackendModule.Type)
   val SPARK_ELASTIC   = ExternalBackendRef(BackendRef(BackendName("spark_elastic")    , BackendCapability.All), sparkcore.fs.elastic.SparkElasticBackendModule.Type)
-  val SPARK_CASSANDRA = ExternalBackendRef(BackendRef(BackendName("spark_cassandra")  , BackendCapability.All), sparkcore.fs.cassandra.FsType)
-
+  val SPARK_CASSANDRA = ExternalBackendRef(BackendRef(BackendName("spark_cassandra")  , BackendCapability.All), sparkcore.fs.cassandra.SparkCassandraBackendModule.Type)
 
   lazy val backendRefs: List[ExternalBackendRef] = List(
     COUCHBASE,
